@@ -3,7 +3,7 @@
   "description": "PhpOffice helper classes",
   "keywords": [
     "docx",
-    "template processor"
+    "template processor helper"
   ],
   "type": "contao-bundle",
   "license": "LGPL-3.0+",
@@ -20,8 +20,8 @@
     "source": "https://github.com/markocupic/phpoffice-bundle"
   },
   "require": {
-    "php": "^7.0||^8.0",
-    "contao/core-bundle": "^4.9",
+    "php": "^7.4||^8.0",
+    "contao/core-bundle": "^4.12",
     "phpoffice/phpword": "^0.18"
   },
   "autoload": {
