@@ -1,22 +1,24 @@
 <?php
-/**
- * PhpOffice helper classes
- * Copyright (c) 2008-2019 Marko Cupic
- * @package phpoffice-bundle
- * @author Marko Cupic m.cupic@gmx.ch, 2019
+
+declare(strict_types=1);
+
+/*
+ * This file is part of Php Office Bundle.
+ *
+ * (c) Marko Cupic 2022 <m.cupic@gmx.ch>
+ * @license GPL-3.0-or-later
+ * For the full copyright and license information,
+ * please view the LICENSE file that was distributed with this source code.
  * @link https://github.com/markocupic/phpoffice-bundle
  */
 
 namespace Markocupic\PhpOfficeBundle\ContaoManager;
 
+use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 
-/**
- * Class Plugin
- * @package Markocupic\PhpOfficeBundle\ContaoManager
- */
 class Plugin implements BundlePluginInterface
 {
     /**
@@ -26,7 +28,7 @@ class Plugin implements BundlePluginInterface
     {
         return [
             BundleConfig::create('Markocupic\PhpOfficeBundle\MarkocupicPhpOfficeBundle')
-                ->setLoadAfter(['Contao\CoreBundle\ContaoCoreBundle'])
+                ->setLoadAfter([ContaoCoreBundle::class]),
         ];
     }
 }
