@@ -1,9 +1,14 @@
 <?php
-/**
- * PhpOffice helper classes
- * Copyright (c) 2008-2019 Marko Cupic
- * @package phpoffice-bundle
- * @author Marko Cupic m.cupic@gmx.ch, 2019
+
+declare(strict_types=1);
+
+/*
+ * This file is part of Php Office Bundle.
+ *
+ * (c) Marko Cupic 2022 <m.cupic@gmx.ch>
+ * @license GPL-3.0-or-later
+ * For the full copyright and license information,
+ * please view the LICENSE file that was distributed with this source code.
  * @link https://github.com/markocupic/phpoffice-bundle
  */
 
@@ -11,10 +16,6 @@ namespace Markocupic\PhpOfficeBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-/**
- * Class MarkocupicPhpOfficeBundle
- * @package Markocupic\PhpOfficeBundle
- */
 class MarkocupicPhpOfficeBundle extends Bundle
 {
 }
